chat-ws/ai/requirements.txt
groq>=1.0.0
pydantic>=2.12.5
python-dotenv>=1.2.1
tenacity>=8.2.0
icalendar>=5.0.0
redis[hiredis]>=5.0.0
